requirements.txt
streamlit
numpy
pillow
tensorflow>=2.7.0
scikit-learn
